backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.10.0
pydantic-settings>=2.0.0
pydantic-core>=2.10.0
python-dotenv>=1.0.0
tortoise-orm>=0.22.0
aerich>=0.7.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
